refactor(haip-79): refactor folder ui placement
