+++ app/login/page.tsx
@@ -1,4 +1,5 @@
-import { Layout, LoginForm } from "@/domain/account/ui";
+import { LoginForm } from "@/domain/account/ui/csr";
+import { Layout } from "@/domain/account/ui/ssr";
 
 const Login = () => (
   <Layout>
