+++ domain/account/ui/csr/VerifyAccount/index.ts
