+++ domain/account/ui/csr/index.ts
@@ -1,4 +1,3 @@
 export * from "./LoginForm";
 export * from "./RegisterForm";
-export * from "./Layout";
 export * from "./VerifyAccount";
