+++ domain/account/ui/ssr/Layout/Layout.tsx
