+++ domain/account/ui/ssr/Layout/index.tsx
@@ -1,0 +1,1 @@
+export * from "./Layout";
